requirements.txt
Flask>=2.0.0
pycryptodome>=3.15.0
qrcode>=7.0.0
Pillow>=8.0.0